Version 1.0.1 alexa-client-sdk
Changes in this update
 - Added a fix to the sample app so that the StateSynchronization event is the first that gets sent to AVS.
 - Added a POST_CONNECTED enum to ConnectionStatusObserver.
 - StateSychronizer now automatically sends a StateSynchronization event when it receives a notification that ACL is CONNECTED.
 - Added make install for installing the AVS Device SDK.
 - Added an optional make networkIntegration for integration tests for slow network (only on Linux platforms).
 - Added shutdown management which fully cleans up SDK objects during teardown.
 - Fixed an issue with AudioPlayer barge-in which was preventing subsequent audio from playing.
 - Changed Mediaplayer buffering to reduce stuttering.
 - Known Issues:
   - Connection loss during listening keeps the app in that state even after connection is regained. Pressing ‘s’ unsticks the state.
   - Play/Pause media restarts it from the beginning.
   - SpeechSynthesizer shows wrong UX state during a burst of Speaks.
   - Quitting the sample app while AudioPlayer is playing something causes a segmentation fault.
   - AudioPlayer sending PlaybackPaused during flash briefing.
   - Long Delay playing live stations on iHeartRadio.
   - Teardown warnings at the end of integration tests.

+++ KWD/KittAi/src/KittAiKeyWordDetector.cpp
@@ -159,21 +159,21 @@ bool KittAiKeyWordDetector::init(avsCommon::utils::AudioFormat audioFormat) {
 }
 
 bool KittAiKeyWordDetector::isAudioFormatCompatibleWithKittAi(avsCommon::utils::AudioFormat audioFormat) {
-    if (audioFormat.numChannels != m_kittAiEngine->NumChannels()) {
+    if (audioFormat.numChannels != static_cast<unsigned int>(m_kittAiEngine->NumChannels())) {
         ACSDK_ERROR(LX("isAudioFormatCompatibleWithKittAiFailed")
                 .d("reason", "numChannelsMismatch")
                 .d("kittAiNumChannels", m_kittAiEngine->NumChannels())
                 .d("numChannels", audioFormat.numChannels));
         return false;
     }
-    if (audioFormat.sampleRateHz != m_kittAiEngine->SampleRate()) {
+    if (audioFormat.sampleRateHz != static_cast<unsigned int>(m_kittAiEngine->SampleRate())) {
         ACSDK_ERROR(LX("isAudioFormatCompatibleWithKittAiFailed")
                 .d("reason", "sampleRateMismatch")
                 .d("kittAiSampleRate", m_kittAiEngine->SampleRate())
                 .d("sampleRate", audioFormat.sampleRateHz));
         return false;
     }
-    if (audioFormat.sampleSizeInBits != m_kittAiEngine->BitsPerSample()) {
+    if (audioFormat.sampleSizeInBits != static_cast<unsigned int>(m_kittAiEngine->BitsPerSample())) {
         ACSDK_ERROR(LX("isAudioFormatCompatibleWithKittAiFailed")
                 .d("reason", "sampleSizeInBitsMismatch")
                 .d("kittAiSampleSizeInBits", m_kittAiEngine->BitsPerSample())
